+++ azext_edge/edge/providers/edge_api/__init__.py
@@ -13,7 +13,7 @@ from .deviceregistry import DEVICEREGISTRY_API_V1, DeviceRegistryResourceKinds
 from .dataflow import DATAFLOW_API_V1B1, DataflowResourceKinds
 from .meta import META_API_V1B1, MetaResourceKinds
 from .arccontainerstorage import ARCCONTAINERSTORAGE_API_V1
-from .secretsynccontroller import SECRETSYNC_API_V1, SECRETSTORE_API_V1
+from .secretstore import SECRETSYNC_API_V1, SECRETSTORE_API_V1
 
 __all__ = [
     "ARCCONTAINERSTORAGE_API_V1",
